+++ qgreenland/config/helpers/layers/gravity_anomalies_geoid.py
@@ -1,6 +1,8 @@
 from qgreenland.config.datasets.geoid import geoid
 from qgreenland.config.datasets.gravity_anomalies import gravity_anomalies
-from qgreenland.config.helpers.steps.build_overviews import build_overviews
+from qgreenland.config.helpers.steps.compress_and_add_overviews import (
+    compress_and_add_overviews,
+)
 from qgreenland.config.helpers.steps.decompress import decompress_step
 from qgreenland.config.helpers.steps.warp_and_cut import warp_and_cut
 from qgreenland.config.project import project
@@ -43,9 +45,10 @@ def _make_layer(
                 ],
                 cut_file=project.boundaries['data'].filepath,
             ),
-            *build_overviews(
+            *compress_and_add_overviews(
                 input_file='{input_dir}/' + filename,
                 output_file='{output_dir}/' + filename,
+                dtype_is_float=True,
             ),
         ],
     )