+++ qgreenland/config/layers/Glaciology/Surface elevation change/__settings__.py
@@ -3,7 +3,7 @@ from qgreenland.models.config.layer_group import LayerGroupSettings
 
 settings = LayerGroupSettings(
     order=[
-        'Observations',
         'Errors',
+        'Observations',
     ],
 )