+++ qgreenland/config/layers/Sea ice/Monthly mean concentration (25 km)/September (min monthly extent)/layers.py
@@ -4,7 +4,7 @@ from qgreenland.config.helpers.layers.sea_ice_concentration import (
     CONCENTRATION_STYLE,
     CONCENTRATION_YEARS,
 )
-from qgreenland.config.helpers.steps.build_overviews import build_overviews
+from qgreenland.config.helpers.steps.compress_and_add_overviews import compress_and_add_overviews
 from qgreenland.config.helpers.steps.warp_and_cut import warp_and_cut
 from qgreenland.config.project import project
 from qgreenland.models.config.layer import Layer, LayerInput
@@ -36,9 +36,10 @@ layers = [
                 output_file='{output_dir}/warped_and_cut.tif',
                 cut_file=project.boundaries['background'].filepath,
             ),
-            *build_overviews(
+            *compress_and_add_overviews(
                 input_file='{input_dir}/warped_and_cut.tif',
                 output_file='{output_dir}/overviews.tif',
+                dtype_is_float=False,
             ),
         ],
     ) for year in CONCENTRATION_YEARS