+++ qgreenland/config/layers/Terrain models/BedMachine/bedmachine.py
@@ -1,5 +1,5 @@
 from qgreenland.config.datasets import bedmachine
-from qgreenland.config.helpers.steps.build_overviews import build_overviews
+from qgreenland.config.helpers.steps.compress_and_add_overviews import compress_and_add_overviews
 from qgreenland.config.helpers.steps.warp_and_cut import warp_and_cut
 from qgreenland.models.config.layer import Layer, LayerInput
 
@@ -51,9 +51,10 @@ layers = [
                 output_file='{output_dir}/warped_and_cut.tif',
                 cut_file='{assets_dir}/greenland_rectangle.geojson',
             ),
-            *build_overviews(
+            *compress_and_add_overviews(
                 input_file='{input_dir}/warped_and_cut.tif',
                 output_file='{output_dir}/overviews.tif',
+                dtype_is_float=False if key == 'errbed' else True,
             ),
         ],
     )