+++ qgreenland/config/layers/Terrain models/dms_gtk_topo.py
@@ -1,7 +1,7 @@
 from qgreenland.config.datasets.dms_gtk import (
     danish_agency_for_data_supply_and_efficiency_gtk_topo_map as dataset,
 )
-from qgreenland.config.helpers.steps.build_overviews import build_overviews
+from qgreenland.config.helpers.steps.compress_and_add_overviews import compress_and_add_overviews
 from qgreenland.models.config.layer import Layer, LayerInput
 
 
@@ -20,9 +20,10 @@ dms_gtk_topo = Layer(
         # NOTE: This layer is not reprojected. We had some issues in the distant
         # past reprojecting this layer correctly. Its internal CRS is
         # `EPSG:32624`.
-        *build_overviews(
+        *compress_and_add_overviews(
             input_file='{input_dir}/dms_gtk_topo.tif',
             output_file='{output_dir}/final.tif',
+            dtype_is_float=False,
         ),
     ],
 )